# src/component-service/src/main/java/com/example/ComponentService/DTOs/ComponentDto.java
package com.example.ComponentService.DTOs;

public class ComponentDto {

    private String serialNumber;

    private String type;

    private String description;

    public String getSerialNumber() {
        return serialNumber;
    }

    public void setSerialNumber(String serialNumber) {
        this.serialNumber = serialNumber;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}