# src/vehicle-service/Zartis.OpenTelemetry.VehicleServiceApi/OpenAPIs/api-docs.json
{"openapi":"3.0.3","info":{"title":"Component service","description":"This is the Microservice dealing with Component information","termsOfService":"https://choosealicense.com/licenses/mit/","contact":{"name":"Me :)"},"license":{"name":"MIT License","url":"https://choosealicense.com/licenses/mit/"},"version":"v1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Inferred Url"}],"tags":[{"name":"component-controller","description":"Component Controller"}],"paths":{"/Components":{"get":{"tags":["component-controller"],"summary":"getComponents","operationId":"getComponentsUsingGET","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ComponentDto"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["component-controller"],"summary":"createComponent","operationId":"createComponentUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComponentDto"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComponentDto"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/Components/{serialNumber}":{"get":{"tags":["component-controller"],"summary":"getComponent","operationId":"getComponentUsingGET","parameters":[{"name":"serialNumber","in":"path","description":"serialNumber","required":true,"style":"simple","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComponentDto"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["component-controller"],"summary":"modifyComponent","operationId":"modifyComponentUsingPUT","parameters":[{"name":"serialNumber","in":"path","description":"serialNumber","required":true,"style":"simple","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComponentDto"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComponentDto"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["component-controller"],"summary":"deleteComponent","operationId":"deleteComponentUsingDELETE","parameters":[{"name":"serialNumber","in":"path","description":"serialNumber","required":true,"style":"simple","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComponentDto"}}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"components":{"schemas":{"ComponentDto":{"title":"ComponentDto","type":"object","properties":{"description":{"type":"string"},"serialNumber":{"type":"string"},"type":{"type":"string"}}}}}}